finish css styling for login/signup
