+++ css/main.css
@@ -3,12 +3,22 @@
 * {
 	margin: 0;
 	padding: 0px;
-    font-family: 'Lato', sans-serif;
+    font-family: 'Lato', 'Roboto', sans-serif;
     color: #282828;
 }
 
-h1, h2, h3, h4, h5, h6, h7 {
-    font-family: 'Lato', sans-serif;
+a,
+a:hover {
+    cursor: pointer;
+    color: #282828; 
+}
+
+a.tags {
+    color: #1c94e0;
+}
+
+h1, h2, h3, h4, h5, h6, h7 p {
+    font-family: 'Lato', 'Roboto', sans-serif;
 	font-weight: 300;
 }
 
@@ -309,8 +319,12 @@ h1, h2, h3, h4, h5, h6, h7 {
 
 .footer > h6  {
     display: inline-block;
-    width: 120px;
     font-size: 14px;
+    margin: 15px 10px;
+}
+
+.footer > h6 > a {
+    color: #1c94e0;
 }
 
 @media only screen and (max-width: 767px){
@@ -375,22 +389,6 @@ h1, h2, h3, h4, h5, h6, h7 {
 
     .footer > h6 {
         display: inline-block;
-        width: 100px;
         font-size: 12px;
     }
-}
-
-a,
-a:hover {
-    cursor: pointer;
-    color: #1c94e0; 
-}
-
-a#tags {
-    color: #1c94e0;
-}
-
-a.user {
-    color: #282828;
-    float: bottom;
 }
