+++ landing-page.html
@@ -19,7 +19,7 @@
 
     <body>
 
-        <!--NAVIGATION BAR-->
+        <!--NAVIGATION BAR START-->
         <div class="navbar navbar-default navbar-fixed-top">
             <div class="navbar-brand">
                 <a class= "navbar-brand-logo" href="landing-page.html">
@@ -38,7 +38,7 @@
                         <h3>Come for the <a>#freefood</a>, stay for the people,</h3>
                         <h4 class="lit-line">and see which events are going to be <img class="lit" src="img/really-lit.png" alt="lit">.</h4> 
                     </div>
-                    <a class="btn btn-primary get-started-button" href="home.html">Get Started</a>
+                    <a class="btn get-started-button" href="home.html">Get Started</a>
                 </div>
                 <div class="col-sm-4"> 
                     <div class="text-container">
@@ -59,7 +59,7 @@
                         <h3>Show off the events your community hosts</h3>    
                         <h4>or host your own private event.</h4>                
                     </div>
-                    <a class="btn btn-primary get-started-button" href="home.html">Host an Event</a>
+                    <a class="btn get-started-button" href="home.html">Host an Event</a>
                 </div>
             </div>
